Refactor ID type from Integer to Long across multiple entities

+++ FoodistaBackend/src/main/java/com/foodista/controller/PostController.java
@@ -59,18 +59,18 @@ public class PostController {
             System.out.println("get_all_post_2");
 
             postService.getAll().forEach(post -> {
-            // System.out.println("getBlogId"+blog.getBlogId());
-            // System.out.println("getUser"+blog.getUser().getId());
-                if(post.getPostType().getPostTypeId() == 1){
+                // System.out.println("getBlogId"+blog.getBlogId());
+                // System.out.println("getUser"+blog.getUser().getId());
+                if (post.getPostType().getPostTypeId() == 1) {
                     PostResponse response = new PostResponse(
-                        post.getPostId(),
-                        post.getBlog().getBlogId(),
-                        post.getAuthor(),
-                        post.getPublishDate()+"",
-                        "",
-                        post.getCategory(),
-                        post.getPostType());
-                
+                            post.getPostId(),
+                            post.getBlog().getBlogId(),
+                            post.getAuthor(),
+                            post.getPublishDate() + "",
+                            "",
+                            post.getCategory(),
+                            post.getPostType());
+
                     posts.add(response);
                 }
             });
@@ -88,18 +88,18 @@ public class PostController {
     }
 
     @GetMapping("/user/{user_id}")
-    public ResponseEntity<List<PostResponse>> getAllPostByUser(@RequestHeader(value = "Authorization", required = false) String token, @PathVariable("user_id") Integer userId) {
+    public ResponseEntity<List<PostResponse>> getAllPostByUser(@RequestHeader(value = "Authorization", required = false) String token, @PathVariable("user_id") Long userId) {
         try {
             System.out.println("getByUser_1");
             List<PostResponse> posts = new ArrayList<PostResponse>();
 
             final int[] tokenUserId = {-1};
 
-            if(token != null && !token.isEmpty()){
+            if (token != null && !token.isEmpty()) {
                 token = token.split(" ")[1];
                 //  System.out.println(token);
                 String tmp = jwtService.extractUserId(token);
-                tokenUserId[0]  =  Integer.valueOf(tmp);
+                tokenUserId[0] = Integer.valueOf(tmp);
 
                 // System.out.println(tokenUserId);
                 // System.out.println("getByUser_2");
@@ -111,16 +111,16 @@ public class PostController {
                 // System.out.println("getByUser_3");
                 // System.out.println(tokenUserId == userId);
 
-                if(post.getPostType().getPostTypeId() == 1 || tokenUserId[0] == userId ){
+                if (post.getPostType().getPostTypeId() == 1 || tokenUserId[0] == userId) {
                     PostResponse response = new PostResponse(
-                        post.getPostId(),
-                        post.getBlog().getBlogId(),
-                        post.getAuthor(),
-                        post.getPublishDate()+"",
-                        "",
-                        post.getCategory(),
-                        post.getPostType());
-                
+                            post.getPostId(),
+                            post.getBlog().getBlogId(),
+                            post.getAuthor(),
+                            post.getPublishDate() + "",
+                            "",
+                            post.getCategory(),
+                            post.getPostType());
+
                     posts.add(response);
                 }
             });
@@ -138,18 +138,18 @@ public class PostController {
 
 
     @GetMapping("/blog/{blog_id}")
-    public ResponseEntity<List<PostResponse>> getAllPostByBlogId (@RequestHeader(value = "Authorization", required = false) String token, @PathVariable("blog_id") Integer blogId) {
+    public ResponseEntity<List<PostResponse>> getAllPostByBlogId(@RequestHeader(value = "Authorization", required = false) String token, @PathVariable("blog_id") Long blogId) {
         try {
             System.out.println("getByBlog_1");
             List<PostResponse> posts = new ArrayList<PostResponse>();
 
             final int[] tokenUserId = {-1};
 
-            if(token != null && !token.isEmpty()){
+            if (token != null && !token.isEmpty()) {
                 token = token.split(" ")[1];
                 //  System.out.println(token);
                 String tmp = jwtService.extractUserId(token);
-                tokenUserId[0]  =  Integer.valueOf(tmp);
+                tokenUserId[0] = Integer.valueOf(tmp);
 
                 // System.out.println(tokenUserId);
                 // System.out.println("getByUser_2");
@@ -157,27 +157,27 @@ public class PostController {
             }
             // System.out.println("getByUser_2");
             Optional<Blog> tmpBlog = blogService.getById(blogId);
-            if(!tmpBlog.isPresent()){
+            if (!tmpBlog.isPresent()) {
                 System.out.println("getByBlog_no_blog");
                 return new ResponseEntity<>(HttpStatus.NOT_FOUND);
             }
             Blog blog = tmpBlog.get();
             Long userId = blog.getUser().getId();
-            
+
             postService.getByBlogId(blogId).forEach(post -> {
                 // System.out.println("getByUser_3");
                 // System.out.println(tokenUserId == userId);
 
-                if(post.getPostType().getPostTypeId() == 1 || tokenUserId[0] == userId ){
+                if (post.getPostType().getPostTypeId() == 1 || tokenUserId[0] == userId) {
                     PostResponse response = new PostResponse(
-                        post.getPostId(),
-                        post.getBlog().getBlogId(),
-                        post.getAuthor(),
-                        post.getPublishDate()+"",
-                        "",
-                        post.getCategory(),
-                        post.getPostType());
-                
+                            post.getPostId(),
+                            post.getBlog().getBlogId(),
+                            post.getAuthor(),
+                            post.getPublishDate() + "",
+                            "",
+                            post.getCategory(),
+                            post.getPostType());
+
                     posts.add(response);
                 }
             });
@@ -194,9 +194,8 @@ public class PostController {
     }
 
 
-
     @GetMapping("/{id}")
-    public ResponseEntity<Post> getBlogById(@PathVariable("id") Integer id) {
+    public ResponseEntity<Post> getBlogById(@PathVariable("id") Long id) {
         Optional<Post> post = postService.getById(id);
         //  System.out.println(blog);
 
@@ -217,7 +216,7 @@ public class PostController {
             String tokenUserId = jwtService.extractUserId(token);
 
             System.out.println("createBlog_1");
-            Long user_id = Long.valueOf(tokenUserId) ;
+            Long user_id = Long.valueOf(tokenUserId);
             Optional<User> user = userRepository.findById(user_id);
 
             System.out.println("createBlog_get_User");
@@ -227,50 +226,50 @@ public class PostController {
             String author = user.get().getFullName();
 
             System.out.println("createBlog_get_blog");
-            Integer blog_id = newpost.getBlog_id();
+            Long blog_id = newpost.getBlog_id();
             Blog blog = new Blog();
-            if(blog_id == null){
+            if (blog_id == null) {
                 System.out.println("createBlog_find_or_create_default_blog");
                 blog = blogService.findOrCreateDeflautBlog(user.get());
-            }else{
+            } else {
                 System.out.println("createBlog_find_blog");
                 Optional<Blog> tmpBlog = blogService.getById(blog_id);
-                if(!tmpBlog.isPresent()){
+                if (!tmpBlog.isPresent()) {
                     System.out.println("createBlog_no_blog");
                     return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-                }else{
+                } else {
                     blog = tmpBlog.get();
-                    if(blog.getUser().getId() != user_id){
+                    if (blog.getUser().getId() != user_id) {
                         System.out.println("createBlog_the_blog_user_not_match");
                         return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
                     }
                 }
-            } 
+            }
 
             System.out.println("createBlog_get_catergory");
             Long category_id = Long.valueOf(newpost.getCategory_id());
 
-            CategoryDetail category =  new CategoryDetail();
+            CategoryDetail category = new CategoryDetail();
             Optional<CategoryDetail> tmpCategory = categoryDetailRepository.findById(category_id);
 
             if (!tmpCategory.isPresent()) {
                 System.out.println("createBlog_no_catergory");
                 return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
-            }else{
+            } else {
                 category = tmpCategory.get();
             }
-            
+
             System.out.println("createBlog_get_post_type");
-            Integer post_type_id = newpost.getPost_type_id();
+            Long post_type_id = newpost.getPost_type_id();
 
 
-            PostType postType =  new PostType();
+            PostType postType = new PostType();
             Optional<PostType> tmpPostType = postTypeRepository.findById(post_type_id);
-            
+
             if (!tmpPostType.isPresent()) {
                 System.out.println("createBlog_no_post_type");
                 return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
-            }else{
+            } else {
                 postType = tmpPostType.get();
             }
 
@@ -279,7 +278,7 @@ public class PostController {
 
             System.out.println("createBlog_newPostContent");
             Post newPostContent = Post.builder()
-                    .postContent(newpost.getPost_content().getBytes() )
+                    .postContent(newpost.getPost_content().getBytes())
                     .blog(blog)
                     .category(category)
                     .postType(postType)
@@ -292,18 +291,18 @@ public class PostController {
             if (createdPost == null) {
                 System.out.println("createBlog_no_post");
                 return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
-            }else{
+            } else {
                 PostResponse responsePost = new PostResponse(
-                    createdPost.getPostId(),
-                    blog.getBlogId(),
-                    author,
-                    timestamp+"",
-                    "success",
-                    category,
-                    postType
+                        createdPost.getPostId(),
+                        blog.getBlogId(),
+                        author,
+                        timestamp + "",
+                        "success",
+                        category,
+                        postType
                 );
-                
-                return new ResponseEntity<>( responsePost, HttpStatus.CREATED);
+
+                return new ResponseEntity<>(responsePost, HttpStatus.CREATED);
             }
 
         } catch (Exception e) {
@@ -313,7 +312,7 @@ public class PostController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<PostResponse> updateBlog(@RequestHeader(value = "Authorization", required = true) String token, @PathVariable("id") Integer id, @RequestBody PostRequest updatePost) {
+    public ResponseEntity<PostResponse> updateBlog(@RequestHeader(value = "Authorization", required = true) String token, @PathVariable("id") Long id, @RequestBody PostRequest updatePost) {
         Optional<Post> post = postService.getById(id);
         //  System.out.println(blog);
 
@@ -329,32 +328,32 @@ public class PostController {
             System.out.println(tmpTokenUserId != original_user_id);
 
 
-            if(tmpTokenUserId != original_user_id){
+            if (tmpTokenUserId != original_user_id) {
                 System.out.println("false token");
-                 return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+                return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
             }
             System.out.println("updatePost_Blog");
 
-            Integer newBlogId = updatePost.getBlog_id();
+            Long newBlogId = updatePost.getBlog_id();
             //check blog id is changed
-            if( newBlogId != null && newBlogId > 0){
+            if (newBlogId != null && newBlogId > 0) {
                 Long oldBlogId = response.getBlog().getBlogId();
-                
+
                 Long tmpBlogId = Long.valueOf(newBlogId);
 
 
                 System.out.println("updatePost_Blog");
                 System.out.println(oldBlogId == tmpBlogId);
 
-                if(oldBlogId != tmpBlogId && newBlogId != null){
+                if (oldBlogId != tmpBlogId && newBlogId != null) {
                     System.out.println("updatePost_Blog_1");
                     Optional<Blog> tmpBlog = blogService.getById(newBlogId);
-                    if(!tmpBlog.isPresent()){
+                    if (!tmpBlog.isPresent()) {
                         System.out.println("updatePost_Blog_2");
                         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-                    }else{
+                    } else {
                         response.setBlog(tmpBlog.get());
-                        if(tmpBlog.get().getUser().getId() != original_user_id){
+                        if (tmpBlog.get().getUser().getId() != original_user_id) {
                             System.out.println("createBlog_the_blog_user_not_match");
                             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
                         }
@@ -365,53 +364,53 @@ public class PostController {
 
             //check category id is changed
             Long oldCategoryId = response.getCategory().getCategoryId();
-            Integer newCategoryId = updatePost.getCategory_id();
-            Long tmpCategoryId= Long.valueOf(newCategoryId);
+            Long newCategoryId = updatePost.getCategory_id();
+            Long tmpCategoryId = Long.valueOf(newCategoryId);
 
             System.out.println("updatePost_Category");
             System.out.println(oldCategoryId == tmpCategoryId);
 
 
-            if(oldCategoryId != tmpCategoryId && newCategoryId != null){
+            if (oldCategoryId != tmpCategoryId && newCategoryId != null) {
                 System.out.println("updatePost_Categgory_1");
                 Optional<CategoryDetail> tmpCategory = categoryDetailRepository.findById(newCategoryId);
-                if(!tmpCategory.isPresent()){
+                if (!tmpCategory.isPresent()) {
                     System.out.println("updatePost_Categgory_2");
                     return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-                }else{
+                } else {
                     response.setCategory(tmpCategory.get());
                 }
             }
 
             //check post type id is changed
-            Integer oldPostTypeId = response.getPostType().getPostTypeId();
-            Integer newPostTypeId = updatePost.getPost_type_id();
+            Long oldPostTypeId = response.getPostType().getPostTypeId();
+            Long newPostTypeId = updatePost.getPost_type_id();
 
-            if(oldPostTypeId != newPostTypeId && newPostTypeId != null){
+            if (oldPostTypeId != newPostTypeId && newPostTypeId != null) {
                 System.out.println("updatePost_PostType_1");
                 Optional<PostType> tmpPostType = postTypeRepository.findById(newPostTypeId);
-                if(!tmpPostType.isPresent()){
+                if (!tmpPostType.isPresent()) {
                     System.out.println("updatePost_PostType_2");
                     return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-                }else{
+                } else {
                     response.setPostType(tmpPostType.get());
                 }
             }
 
             response.setPostContent(updatePost.getPost_content().getBytes());
-            
+
             postService.save(response);
 
             PostResponse responsePost = new PostResponse(
-                response.getPostId(),
-                original_user_id,
-                response.getAuthor(),
-                response.getPublishDate()+"",
-                "success",
-                response.getCategory(),
-                response.getPostType()
+                    response.getPostId(),
+                    original_user_id,
+                    response.getAuthor(),
+                    response.getPublishDate() + "",
+                    "success",
+                    response.getCategory(),
+                    response.getPostType()
             );
-           
+
 
             return new ResponseEntity<>(responsePost, HttpStatus.OK);
         } else {
@@ -421,7 +420,7 @@ public class PostController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> deleteBlog(@RequestHeader(value = "Authorization", required = true) String token, @PathVariable("id") Integer id) {
+    public ResponseEntity<HttpStatus> deleteBlog(@RequestHeader(value = "Authorization", required = true) String token, @PathVariable("id") Long id) {
 
         try {
             Optional<Post> post = postService.getById(id);
@@ -438,19 +437,19 @@ public class PostController {
                 System.out.println(tmpTokenUserId != original_user_id);
 
 
-                if(tmpTokenUserId != original_user_id){
+                if (tmpTokenUserId != original_user_id) {
                     return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
                 }
 
 
-                Integer archivePostId = 3;
+                Long archivePostId = 3L;
 
                 Optional<PostType> tmpPostType = postTypeRepository.findById(archivePostId);
                 response.setPostType(tmpPostType.get());
 
                 postService.save(response);
                 return new ResponseEntity<>(HttpStatus.OK);
-            }else{
+            } else {
                 return new ResponseEntity<>(HttpStatus.NOT_FOUND);
             }
         } catch (Exception e) {
