+++ FoodistaBackend/src/main/java/com/foodista/repositories/PostTypeRepository.java
@@ -8,11 +8,11 @@ import org.springframework.data.jpa.repository.Query;
 import com.foodista.entities.PostType;
 
 
-public interface PostTypeRepository extends JpaRepository<PostType, Integer> {
-    
+public interface PostTypeRepository extends JpaRepository<PostType, Long> {
+
     @Query("SELECT pt FROM PostType pt")
     Iterable<PostType> findAllPostType();
 
     @Query("SELECT pt FROM PostType pt WHERE pt.postTypeId = ?1")
-    Optional<PostType> findById(Integer id);
+    Optional<PostType> findById(Long id);
 }
